+++ compiler-core/src/error.rs
@@ -3290,13 +3290,13 @@ being used as a type constructor."
             }
         }
 
-        TypeError::UnnecessarySpreadOperator { location, arity } => {
+        TypeError::RecordUpdateWithoutChanges { location, arity } => {
             let text = wrap_format!(
                 "This record has {arity} fields and you have already \
 assigned variables to all of them."
             );
             Diagnostic {
-                title: "Unnecessary spread operator".into(),
+                title: "Record update without changes".into(),
                 text,
                 hint: None,
                 level: Level::Error,
@@ -5077,6 +5077,25 @@ Be sure to finish it before running your program.",
             location: Some(Location {
                 label: Label {
                     text: Some("This code is incomplete".into()),
+                    span: *location,
+                },
+                path: path.clone(),
+                src: src.clone(),
+                extra_labels: vec![],
+            }),
+        },
+
+        TypeError::ListPrependWithoutElements {
+            location,
+            tail_location: _,
+        } => Diagnostic {
+            title: "Redundant list prepend".to_string(),
+            text: "See: https://tour.gleam.run/basics/lists/".to_string(),
+            hint: Some("`list` can be used directly instead of `[..list]`".into()),
+            level: Level::Error,
+            location: Some(Location {
+                label: Label {
+                    text: Some("This prepend does nothing".into()),
                     span: *location,
                 },
                 path: path.clone(),