+++ compiler-core/src/type_/error.rs
@@ -242,7 +242,7 @@ pub enum Error {
         reason: UnsafeRecordUpdateReason,
     },
 
-    UnnecessarySpreadOperator {
+    RecordUpdateWithoutChanges {
         location: SrcSpan,
         arity: usize,
     },
@@ -708,6 +708,28 @@ pub enum Error {
     RecordUpdateVariantWithNoFields {
         location: SrcSpan,
     },
+
+    /// Redundantly prepending a list:
+    ///
+    /// ```gleam
+    /// const wibble = [1, 2]
+    ///
+    /// const wobble = [..wibble]
+    ///              // ^^^^^^^^ This prepending is redundant!
+    /// ```
+    ///
+    ListPrependWithoutElements {
+        location: SrcSpan,
+        /// We hold onto the location of the tail expression for the code action.
+        ///
+        /// ```gleam
+        /// const wibble = [..wobble]
+        ///                   ^^^^^^ here
+        /// ```
+        ///
+        tail_location: SrcSpan,
+    },
+
     /// When a constant contains a todo.
     /// Unlike todo _expressions_, todo constants are a compile time error: we
     /// want the developer to take care of them before they can run their code.
@@ -1336,7 +1358,7 @@ impl Error {
             | Error::UnknownRecordField { location, .. }
             | Error::IncorrectArity { location, .. }
             | Error::UnsafeRecordUpdate { location, .. }
-            | Error::UnnecessarySpreadOperator { location, .. }
+            | Error::RecordUpdateWithoutChanges { location, .. }
             | Error::IncorrectTypeArity { location, .. }
             | Error::TypeUsedAsAConstructor { location, .. }
             | Error::CouldNotUnify { location, .. }
@@ -1401,7 +1423,11 @@ impl Error {
             | Error::RecordUpdateVariantWithNoFields { location }
             | Error::QualifiedTypeMissingName { location }
             | Error::TodoConstant { location }
-            | Error::LowercaseBoolPattern { location } => location.start,
+            | Error::LowercaseBoolPattern { location }
+            | Error::ListPrependWithoutElements {
+                location,
+                tail_location: _,
+            } => location.start,
             Error::UnknownLabels { unknown, .. } => {
                 unknown.iter().map(|(_, s)| s.start).min().unwrap_or(0)
             }