+++ compiler-core/src/type_/tests.rs
@@ -3656,3 +3656,26 @@ pub const full_list = [1.0, 2.0, ..list]
 "
     );
 }
+
+#[test]
+fn redundant_list_prepend_variable() {
+    assert_module_error!(
+        "fn main() {
+  let wibble = [1, 2]
+  let wobble = [..wibble]
+}"
+    );
+}
+
+#[test]
+fn redundant_list_prepend_constant() {
+    assert_module_error!(
+        "const wibble = [1, 2]
+const wobble = [..wibble]"
+    );
+}
+
+#[test]
+fn redundant_list_prepend_nested() {
+    assert_module_error!("const wibble = [..[..[..[..[1, 2]]]]]");
+}